# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateProject": {
        "properties": {
          "command_id": {
            "description": "The idempotency key (spec §3.2). The same id with the same request\nreplays the first answer; with a different one it is a conflict.",
            "type": "string"
          },
          "directory": {
            "description": "Absolute path to an existing directory, in any spelling; it is stored\nand returned canonical.",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "command_id",
          "slug",
          "name",
          "directory"
        ],
        "type": "object"
      },
      "CreateThread": {
        "properties": {
          "command_id": {
            "description": "The idempotency key (spec §3.2), scoped to the project.",
            "type": "string"
          },
          "harness": {
            "description": "The CLI for this conversation; `claude-code` when omitted.",
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "command_id",
          "title"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "The body of every error this API answers: the stable code a client\nmatches on (spec §3.4), and a human message nobody should match on. A\nrequest axum rejects before a handler runs (malformed JSON, a missing query\nparameter) gets one too, through\n[`rejections_as_error_bodies`].",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "current_revision": {
            "description": "`REVISION_CONFLICT` only: the resource's revision now, to read again at.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "problems": {
            "description": "`WORKFLOW_VALIDATION_FAILED` only: each problem, the same sentences\n`message` joins.",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ErrorCode": {
        "description": "Stable codes clients pattern-match on. Never match on human text.\nSpec §3.4. `Blocked`/`Rejected` are domain outcomes and never appear here.",
        "enum": [
          "HARNESS_START_FAILED",
          "PROCESS_SPAWN_FAILED",
          "PROCESS_TERMINATED",
          "PROCESS_TERMINATION_FAILED",
          "RUNTIME_STOPPING",
          "STORAGE_UNAVAILABLE",
          "STORAGE_MIGRATION_FAILED",
          "STORAGE_CONSTRAINT_VIOLATION",
          "COMMAND_CONFLICT",
          "IDEMPOTENCY_KEY_REQUIRED",
          "INVALID_COMMAND",
          "INVALID_CURSOR",
          "AGENT_AUTH_FAILED",
          "AGENT_UNSUPPORTED_PROFILE",
          "PATH_INVALID",
          "PATH_NOT_FOUND",
          "PATH_NOT_A_DIRECTORY",
          "PATH_ACCESS_DENIED",
          "PATH_ALREADY_EXISTS",
          "PATH_UNAVAILABLE",
          "ORIGIN_REFUSED",
          "HARNESS_UNAVAILABLE",
          "SETTING_NOT_OFFERED",
          "MODE_NOT_ALLOWED",
          "HARNESS_LOCKED",
          "THREAD_BUSY",
          "FORK_POINT_NOT_SUPPORTED",
          "WORKFLOW_FROZEN_IMMUTABLE",
          "WORKFLOW_VALIDATION_FAILED",
          "REVISION_CONFLICT",
          "PROJECT_HAS_THREADS",
          "GRANT_SCOPE",
          "GRANT_INVALID"
        ],
        "type": "string"
      },
...
      "PlanningThread": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "forked_from_thread": {
            "description": "The thread this one was forked from, if it is a fork (spec §12.9).",
            "type": [
              "string",
              "null"
            ]
          },
          "harness": {
            "description": "The CLI this conversation runs on (`agent::policy`): chosen at\ncreation, changeable until the thread's first operation, then fixed\n(spec §12.6).",
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/ThreadId"
          },
          "project_id": {
            "$ref": "#/components/schemas/ProjectId"
          },
          "removed_at": {
            "description": "Set when this conversation was removed; its history remains readable.",
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "title",
          "status",
          "created_at",
          "removed_at",
          "harness",
          "forked_from_thread"
        ],
        "type": "object"
      },
...
      "Project": {
        "properties": {
          "allowed_modes": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "description": "Per harness kind, the modes this project's turns may use (spec §12.5).\nEach list is a set: its order carries no meaning. Every known harness\nis present; an empty list means no turn can start on it.",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "created_at": {
            "type": "string"
          },
          "directory": {
            "description": "Spec §4.2: the directory this project's turns run in, as\n[`ProjectDirectory`] resolved it. `None` only for a project created\nbefore projects owned one (`migrations/0003_project_directory.sql`);\na turn on such a project fails at Prepare rather than running in the\ndaemon's own working directory.",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/ProjectId"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "created_at",
          "allowed_modes"
        ],
        "type": "object"
      },
      "ProjectId": {
        "format": "uuid",
        "type": "string"
      },
...
      "ThreadId": {
        "format": "uuid",
        "type": "string"
      },
...
      "UpdateProject": {
        "description": "Sets the modes this project allows, per harness (spec §12.5).",
        "properties": {
          "allowed_modes": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "description": "Per harness kind, the modes this project allows; a harness not named\nkeeps its modes. Only modes in Shadows' policy for that harness.",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "command_id": {
            "description": "The idempotency key (spec §3.2), scoped to the project.",
            "type": "string"
          }
        },
        "required": [
          "command_id",
          "allowed_modes"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The local daemon's HTTP API. Every client is on another origin; only origins passed with `--allow-origin` may call it. There is no authentication: the daemon binds to loopback (spec §1, OPEN block on remote access). Errors answered by the API carry `ErrorBody`, whose `code` is stable; a request axum rejects before a handler runs (malformed JSON, a missing parameter) is answered in plain text.",
    "license": {
      "name": ""
    },
    "title": "shadows",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects": {
      "get": {
        "operationId": "list_projects",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "STORAGE_UNAVAILABLE"
          }
        },
        "summary": "Every project, oldest first.",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Created, or the replay of the same command"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PATH_INVALID, PATH_NOT_A_DIRECTORY"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PATH_ACCESS_DENIED"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PATH_NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "COMMAND_CONFLICT, or STORAGE_CONSTRAINT_VIOLATION: the slug is in use"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "STORAGE_UNAVAILABLE, PATH_UNAVAILABLE"
          }
        },
        "summary": "Creates a project owning an existing directory.",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{id}": {
      "delete": {
        "operationId": "remove_project",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectId"
            }
          },
          {
            "description": "The idempotency key (spec §3.2), scoped to the project.",
            "in": "query",
            "name": "command_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Removed, or the replay of the same command"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "INVALID_COMMAND: no `command_id`"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "INVALID_COMMAND: no such project"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_HAS_THREADS, or COMMAND_CONFLICT"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "STORAGE_UNAVAILABLE"
          }
        },
        "summary": "Removes a project that holds no planning thread (spec §4.2): it is listed\nnowhere again, its code index, its links and its MCP grants go, and its\nslug stays taken. A replay answers the removed project.",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "update_project",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "INVALID_COMMAND: no such project"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "COMMAND_CONFLICT"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "SETTING_NOT_OFFERED: a mode outside Shadows' policy"
          }
        },
        "summary": "Sets the modes this project allows, per harness (spec §12.5). A turn's\nmode is checked against them when it starts. A replay answers the project\nas it now stands.",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/projects/{id}/threads": {
      "get": {
        "operationId": "list_threads",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanningThread"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "STORAGE_UNAVAILABLE"
          }
        },
        "summary": "A project's planning threads, oldest first. An unknown project has none.",
        "tags": [
          "threads"
        ]
      },
      "post": {
        "operationId": "create_thread",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThread"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningThread"
                }
              }
            },
            "description": "Created, or the replay of the same command"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "INVALID_COMMAND: no such project"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "COMMAND_CONFLICT"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "SETTING_NOT_OFFERED: a harness Shadows does not know"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "STORAGE_UNAVAILABLE"
          }
        },
        "summary": "Creates a planning thread in a project.",
        "tags": [
          "threads"
        ]
      }
    },
...
  },
...
}